+++ src/constants.js
@@ -1,5 +1,10 @@
+// Global State Keys
+
 module.exports = {
-    // Global State Keys
+    // Account Data
+    MEERKATIO_TOKEN: "meerkatio_token",
+
+    // Notification Data
     NOTIF_COUNT_KEY: "notif_count",
     NOTIF_TOTAL_DURATION_KEY: "notif_total_duration",
     NOTIF_HISTORY: "notif_history"